Updates styling and layout in SucursalCard and PageLayout components
Improves the styling of the SucursalCard component by modifying the classNames in the article element to add the 'z-10' class when the showModal state is true, ensuring the element is displayed on top.

In the PageLayout component, the 'mt-[160px]' class in the main element is changed to 'mt-[100px]' to adjust the top margin.

These changes enhance the visual presentation and user experience of the components.